Added separate scans for normal and hidden truecrypt volume lookups
TrueCrypt4 seems to have a bug: Sometimes the hidden header seems to get
overwritten for no apparent cause.

I'm afraid it might get overwritten with normal user data in that case,
so it seems unwise to restore the (unused) hidden header in such cases,
because that would overwrite the user data as well.

The current TC Volume Lookup code uses a checksum of both the normal
and the hidden header in order to identifiy a volume. If the hidden
header becomes overwritten, the header lookup will be unable to match
the volume.

Therefore, the new script works in two phases: First, it only tries to
match the normal headers. If no match was found, a second pass will be
run matching the hidden headers only.

This should work with all normal volumes, and also with those hidden
volumes where the hidden headers have not been damaged.

Another change of this commit is the xauth "stealer" script. It now
supports to import the "stolen" authority not only to the current user,
but also to specific different user. The default destination is still
the current user, though.