Added script for swapping /usr/share/mime/mime.cache
This file is a pain, because it gets updated more often than necessary,
and then many already-running hold references to the previous copy of
the file, thus disabling remounting /usr as read-only.

This script saves the original mime.cache file with a new name, creating
a copy which will be replaced by subsequent updates.

When run again the script compares the updated and old versions,
restoring the old version if no differences are found.

/usr can then be remounted read-only again without a need to restart
all the applications using mime.cache first.