Support mixed targets in dmsetup tabled as well
Enhance the script for displaying the contents of dmsetup table files so
that individual files can contain a mix of "crypt" targets and other
targets, but all "crypt" keys are still filtered out.

To achieve this, move the "crypt" detection into the "sed" script
itself, rather than letting the outer shell script do it.