Add live counters from interface statistics
But only do this if no arguments are given.

Otherwise, assume historic information is queried for, and only use the
syslog data of already terminated sessions.

Also support a day count of 0 days before the current one, meaning the
live counters will not be included.

Remove the "-p" option for terminating the connection internally. It is
now no longer needed, because the live counters can provide the
previously missing information.