Change the UID for VFAT mounts
It used to be "operator" because that sounded just right, but
Debian-based installations do not seem to define such a system user any
longer. Falling back to "root". (Maybe "sys" would be a more appropriate
user, but I am not sure for what this system account is actually good
for.)