Add "noatime" for FAT mount helper script
Usually no-one needs the "atime" on a VFAT volume, especially as such
volumes are most likely flash-based devices nowadays.

Also get rid of the "vfatflashmount"-Script, because it is hopelessly
outdated. When it was conceived a long time ago, the kernel hat a
different I/O scheduler which seemed more appropriate than the default
scheduler, and so that script set it. However, that scheduler does no
longer exist in newer kernels, and so there is virtually no useful
difference between both VFAT mount helper scripts any more.