Add support for appending duplicated sections
The new "duplicate_append"-key works like "duplicate" but appends the
section to the end of the file being customized.

This is implemented using a temporary file.

Add file-type support for Scala and Julia.

Add a file-contents sniffing trigger for Python files which are embedded
within shell scripts.

At least in the current joe version 4.4.1 there it a problem with
matching \[...] sets. It just does not seem to work as part of the
sniffing patterns, even though it works fine when used interactively.