Update most REs to new syntax
Now instead of \+x and \[...] one has to write x+ and [...],
respectively.

Also \* became a postix operator and is no longer a shell glob matching
arbitrary characters.

Heaven knows what else has changes.

Update most of the patterns, but not all of them. The rest will follow
on an as-needed basis.