Disable new options because they don't work
Neither the "noatime", "relatime" nor "ro"-options seem to work when
embedded into the superblock via '-E mount_opts=$OPTIONLIST'.

When mounting, the kernel then always logs an error message such as
'failed to parse options in superblock: noatime' and ignores the option.

Until this is resolved, disabling the new options for now.