Implement "write", add "encrypt" and "decrypt"
Change the signature of the "write" command of the "afis" utility so
that it only affects a single stripe store rather than updating a
primary and backup store in the same operation.

Also add a wrapper for "gbcrypt2" which combines a user-provided
password file with a synthetic password derived from a specifiy AFIS
slot.

The new "encrypt" end "decrypt" command use the combined password to
encrypt store in the slot-area of the AFIS directory, or decrypt its
contents.

The decrypted contents are usually a master password used for some
encrypted storage device or encrypted storage image file.