Add a loader for stacked dmcrypt tables
This allows cascading ciphers like in TrueCrypt with dmcrypt.

In order to use this, create multiple dmcrypt tables, starting with the
outermost one. Each following table shall be stacked on the previous
one.

Then put all those table definitions into a file and precede each table
with the name it shall be mapped as.

Then feed this file as standard input into script
"dmcrypt-multitable-loader" and it will map all the tables.

It is also possible to specify option "-u". In this case, the stacked
tables will be unmapped rather than being mapped.